+++ packages/client/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @richie-rpc/client
 
+## 1.2.2
+
+### Patch Changes
+
+- add AbortController and jsonSchemaParams support
+
 ## 1.2.1
 
 ### Patch Changes