+++ packages/openapi/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @richie-rpc/openapi
 
+## 1.1.2
+
+### Patch Changes
+
+- add AbortController and jsonSchemaParams support
+
 ## 1.1.1
 
 ### Patch Changes