+++ packages/react-query/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @richie-rpc/react-query
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @richie-rpc/client@1.2.2
+
 ## 1.0.2
 
 ### Patch Changes